requirements.txt
python-telegram-bot
requests
urllib2